Make it reusable component for create, and edit page.

+++ src/main/java/com/demo/struts1/model/Spotify.java
@@ -4,17 +4,17 @@ import java.math.BigInteger;
 
 public class Spotify {
 
-    private BigInteger id;
+    private Long id;
     private String trackName;
     private String artistName;
     private String genre;
     private Integer popularity;
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
